# api/openapi/v1/openapi.yaml
openapi: 3.0.0
info:
  title: CodeGrapher Browser API
  version: 1.0.0
...
paths:
  /repositories:
    get:
      operationId: listRepositories
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /repositories/{repositoryId}:
    get:
      operationId: getRepository
      parameters:
        - name: repositoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositorySummary'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /repositories/{repositoryId}/revisions/{revision}/files:
    get:
      operationId: getFile
      parameters:
        - name: repositoryId
          in: path
          required: true
          schema:
            type: string
        - name: revision
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /repositories/{repositoryId}/revisions/{revision}/search:
    get:
      operationId: searchSymbols
      parameters:
        - name: repositoryId
          in: path
          required: true
          schema:
            type: string
        - name: revision
          in: path
          required: true
          schema:
            type: string
        - name: query
          in: query
          required: true
          schema:
            type: string
          explode: false
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /repositories/{repositoryId}/revisions/{revision}/symbols/{symbolId}:
    get:
      operationId: getSymbol
      parameters:
        - name: repositoryId
          in: path
          required: true
          schema:
            type: string
        - name: revision
          in: path
          required: true
          schema:
            type: string
        - name: symbolId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /repositories/{repositoryId}/revisions/{revision}/symbols/{symbolId}/graph:
    get:
      operationId: getSymbolGraph
      parameters:
        - name: repositoryId
          in: path
          required: true
          schema:
            type: string
        - name: revision
          in: path
          required: true
          schema:
            type: string
        - name: symbolId
          in: path
          required: true
          schema:
            type: string
        - name: direction
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GraphDirection'
          explode: false
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: maxNodes
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: maxEdges
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
security:
  - BearerAuth: []
components:
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
...
    FileResponse:
      type: object
      required:
        - repositoryId
        - revision
        - path
        - language
        - size
        - lineCount
        - contentHash
        - content
        - freshness
      properties:
        repositoryId:
          type: string
        revision:
          type: string
        path:
          type: string
        language:
          type: string
        size:
          type: integer
          format: int64
        lineCount:
          type: integer
          format: int32
        contentHash:
          type: string
        content:
          type: string
        freshness:
          $ref: '#/components/schemas/Freshness'
    Freshness:
      type: object
      required:
        - state
      properties:
        state:
          $ref: '#/components/schemas/FreshnessState'
        indexedAt:
          type: string
          format: date-time
        lastSuccessfulSync:
          type: string
          format: date-time
        lastError:
          type: string
    FreshnessState:
      type: string
      enum:
        - ready
        - updating
        - stale
    GraphDirection:
      type: string
      enum:
        - incoming
        - outgoing
        - both
...
    RepositoryList:
      type: object
      required:
        - repositories
      properties:
        repositories:
          type: array
          items:
            $ref: '#/components/schemas/RepositorySummary'
    RepositorySummary:
      type: object
      required:
        - id
        - label
        - freshness
        - fileCount
        - symbolCount
        - edgeCount
      properties:
        id:
          type: string
        label:
          type: string
        remote:
          type: string
        branch:
          type: string
        headCommit:
          type: string
        revision:
          type: string
        freshness:
          $ref: '#/components/schemas/Freshness'
        fileCount:
          type: integer
          format: int32
        symbolCount:
          type: integer
          format: int32
        edgeCount:
          type: integer
          format: int32
    SearchResponse:
      type: object
      required:
        - repositoryId
        - revision
        - query
        - results
        - limit
        - truncated
        - freshness
      properties:
        repositoryId:
          type: string
        revision:
          type: string
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        limit:
          type: integer
          format: int32
        truncated:
          type: boolean
        freshness:
          $ref: '#/components/schemas/Freshness'
...
    SymbolResponse:
      type: object
      required:
        - repositoryId
        - revision
        - symbol
        - freshness
      properties:
        repositoryId:
          type: string
        revision:
          type: string
        symbol:
          $ref: '#/components/schemas/Symbol'
        freshness:
          $ref: '#/components/schemas/Freshness'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://127.0.0.1:7331/codegrapher/v1
    description: CodeGrapher daemon browser API
    variables: {}